+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - `%c%` internal helper now uses `data.table::fifelse()` to avoid falsely converting dates to integers. (#214)
 - `load_schedules()` cleans the `roof` variable in order to avoid nflverse model issues. (#218)
+- `join_coalesce()` coerces x/y args to data.frame and will return a data.frame
 
 ---